protege operações para que apenas a pessoa que criou a task possa manipular

+++ api/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.response import Response
+from rest_framework.exceptions import PermissionDenied
 from .serializers import TaskSerializer, UserSerializer
 from .models import Task
 
@@ -23,7 +24,7 @@ def apiOverview(request):
 @api_view(['GET'])
 @permission_classes((IsAuthenticated,))
 def taskList(request, format=None):
-    tasks = Task.objects.all()
+    tasks = Task.objects.filter(author=request.user.username)
     serializer = TaskSerializer(tasks, many=True)
     return Response(serializer.data)
 
@@ -31,12 +32,15 @@ def taskList(request, format=None):
 @permission_classes((IsAuthenticated,))
 def taskDetail(request, id):
     task = Task.objects.get(id=id)
+    if task.author != request.user.username:
+        raise PermissionDenied({"message":"You don't have permission to do this"})
     serializer = TaskSerializer(task, many=False)
     return Response(serializer.data)
 
 @api_view(['POST'])
 @permission_classes((IsAuthenticated,))
 def createTask(request):
+    request.data.update({"author": request.user.username})
     serializer = TaskSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -46,6 +50,8 @@ def createTask(request):
 @permission_classes((IsAuthenticated,))
 def updateTask(request, id):
     task = Task.objects.get(id=id)
+    if task.author != request.user.username:
+        raise PermissionDenied({"message":"You don't have permission to do this"})
     serializer = TaskSerializer(instance=task, data=request.data, partial=True)
     if serializer.is_valid():
         serializer.save()
@@ -55,5 +61,7 @@ def updateTask(request, id):
 @permission_classes((IsAuthenticated,))
 def deleteTask(request, id):
     task = Task.objects.get(id=id)
+    if task.author != request.user.username:
+        raise PermissionDenied({"message":"You don't have permission to do this"})
     task.delete()
     return Response("Task deleted")
